move personIdentityQuery
