+++ frontend/src/employee-frontend/App.tsx
@@ -21,7 +21,6 @@ import EmployeeRoute from './components/EmployeeRoute'
 import FinancePage from './components/FinancePage'
 import { Footer } from './components/Footer'
 import Header from './components/Header'
-import IncomeStatementPage from './components/income-statements/IncomeStatementPage'
 import LoginPage from './components/LoginPage'
 import MobilePairingModal from './components/MobilePairingModal'
 import PersonalMobileDevicesPage from './components/PersonalMobileDevicesPage'
@@ -57,6 +56,7 @@ import HolidayAndTermPeriodsPage from './components/holiday-term-periods/Holiday
 import HolidayPeriodEditor from './components/holiday-term-periods/HolidayPeriodEditor'
 import PreschoolTermPeriodEditor from './components/holiday-term-periods/PreschoolTermPeriodEditor'
 import QuestionnaireEditor from './components/holiday-term-periods/QuestionnaireEditor'
+import IncomeStatementPage from './components/income-statements/IncomeStatementPage'
 import IncomeStatementsPage from './components/income-statements/IncomeStatementsPage'
 import InvoicePage from './components/invoice/InvoicePage'
 import InvoicesPage from './components/invoices/InvoicesPage'
